requirements.txt
Flask==2.0.0
gunicorn==20.1.0
beautifulsoup4==4.9.1
line-bot-sdk==1.19.0
imgurpython==1.1.7
twstock~=1.3.1
pandas~=1.2.4
pymongo~=3.11.4
matplotlib~=3.4.2
lxml
Pillow~=8.2.0
selenium~=4.15.2
requests~=2.25.1
